Split worker spawning out of Playground CLI runner
