+++ packages/playground/cli/src/blueprints-v2/worker-thread-v2.ts
@@ -47,7 +47,8 @@ import { existsSync } from 'fs';
 import path from 'path';
 import { rootCertificates } from 'tls';
 import { MessageChannel, type MessagePort, parentPort } from 'worker_threads';
-import { type RunCLIArgs, spawnWorkerThread } from '../run-cli';
+import { type RunCLIArgs } from '../run-cli';
+import { spawnWorkerThread } from '../spawn-worker-thread';
 import type {
 	PhpIniOptions,
 	PHPInstanceCreatedHook,
